Refactor: 메인 페이지 컴포넌트 분리
1. 기존의 사이드바 / 메인 컨테이너로 구분되던 컴포넌트 구조를
2. 사이드바 / 서비스정보 / 방 리스트 로 3개의 컴포넌트로 분리하였습니다.
