+++ client/src/components/RoomBox.tsx
@@ -1,6 +1,6 @@
 import { Link } from 'react-router-dom';
 import styled, { css } from 'styled-components';
-import { RoomInfo } from '@pages/Main/Main';
+import { RoomInfo } from './main/RoomList';
 
 const ROOM_WIDTH = 20;
 const ROOM_HEIGHT = ROOM_WIDTH * 0.75;
